requirements.txt
pandas==1.3.2
Pillow==8.3.2
requests~=2.26.0
